Extract string utilities to own file
