+++ src/celengine/constellation.cpp
@@ -7,7 +7,7 @@
 // as published by the Free Software Foundation; either version 2
 // of the License, or (at your option) any later version.
 
-#include <celutil/util.h>
+#include <celutil/stringutils.h>
 #include "constellation.h"
 
 using namespace std;
