+++ src/celestia/celestiacore.cpp
@@ -38,6 +38,7 @@
 #include <celutil/debug.h>
 #include <celutil/gettext.h>
 #include <celutil/utf8.h>
+#include <celutil/util.h>
 #include <celcompat/filesystem.h>
 #include <Eigen/Geometry>
 #include <iostream>
