+++ src/celutil/filetype.cpp
@@ -7,9 +7,7 @@
 // as published by the Free Software Foundation; either version 2
 // of the License, or (at your option) any later version.
 
-#include <cctype>
-#include <cstdlib>
-#include "util.h"
+#include "stringutils.h"
 #include "filetype.h"
 
 using namespace std;
