+++ src/celutil/util.cpp
@@ -10,7 +10,7 @@
 // of the License, or (at your option) any later version.
 
 #include <config.h>
-#include <celutil/debug.h>
+#include <fmt/printf.h>
 #include "util.h"
 #include "gettext.h"
 #ifdef _WIN32
@@ -25,48 +25,6 @@
 #endif // !_WIN32
 
 using namespace std;
-
-
-int compareIgnoringCase(const string& s1, const string& s2)
-{
-    string::const_iterator i1 = s1.begin();
-    string::const_iterator i2 = s2.begin();
-
-    while (i1 != s1.end() && i2 != s2.end())
-    {
-        if (toupper(*i1) != toupper(*i2))
-            return (toupper(*i1) < toupper(*i2)) ? -1 : 1;
-        ++i1;
-        ++i2;
-    }
-
-    return s2.size() - s1.size();
-}
-
-
-int compareIgnoringCase(const string& s1, const string& s2, int n)
-{
-    string::const_iterator i1 = s1.begin();
-    string::const_iterator i2 = s2.begin();
-
-    while (i1 != s1.end() && i2 != s2.end() && n > 0)
-    {
-        if (toupper(*i1) != toupper(*i2))
-            return (toupper(*i1) < toupper(*i2)) ? -1 : 1;
-        ++i1;
-        ++i2;
-        n--;
-    }
-
-    return n > 0 ? s2.size() - s1.size() : 0;
-}
-
-
-bool CompareIgnoringCasePredicate::operator()(const string& s1,
-                                              const string& s2) const
-{
-    return compareIgnoringCase(s1, s2) < 0;
-}
 
 
 fs::path LocaleFilename(const fs::path &p)
@@ -172,7 +130,7 @@ fs::path homeDir()
     return fs::path();
 }
 
-bool GetTZInfo(std::string &tzName, int &dstBias)
+bool GetTZInfo(std::string_view tzName, int &dstBias)
 {
 #ifdef _WIN32
     TIME_ZONE_INFORMATION tzi;
